+++ rainmapper-app/app/rainmapper_core/sources/sodapy_local/__init__.py
@@ -1,7 +1,0 @@
-from rainmapper_core.sources.sodapy_local.socrata import Socrata
-from rainmapper_core.sources.sodapy_local import version
-
-__all__ = [
-    "Socrata",
-]
-__version__ = version.__version__